Progreso en Evaluaciones Pendientes

+++ everisapi.API/Controllers/EvaluacionController.cs
@@ -477,8 +477,8 @@ namespace everisapi.API.Controllers
       try
       {
         //Recoge si existe la evaluación si es asi la devuelve si no es así muestra un error
-        int? progress = _evaluacionInfoRepository.CalculateEvaluationProgress(idEvaluacion, idAssessment);
-
+        float? progress = _evaluacionInfoRepository.CalculateEvaluationProgress(idEvaluacion, idAssessment);
+        
         if (progress == null)
         {
           _logger.LogInformation("La evaluación información con id " + idEvaluacion + " no pudo ser encontrado.");
